logout admin if token is expired

+++ nwt-front/nwt-election/src/components/admin/js/Users.js
@@ -3,35 +3,40 @@ import axios from "axios";
 import '../css/Users.css';
 import { useHistory } from "react-router-dom";
 import Header from "./Header";
+import {checkExpiration, useHandleLogout} from "../../HelpFunctions";
 
 const Users = () => {
     
     const [users, setUsers] = useState([]); // State to store admin data
+    const handleLogout = useHandleLogout();
 
     const handleDelete = async (email) => {
-        try {
-            const isConfirmed = window.confirm("Are you sure you want to delete this users?");
-            if (!isConfirmed) {
-                return; // Do nothing if the user cancels
-            }
-
-            const BASE_URL = process.env.REACT_APP_BASE_URL ||  'http://localhost:8080';
-            const token = localStorage.getItem('access_token');
-    
-            const response = await fetch(`${BASE_URL}/users/delete-admin?email=${email}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Authorization': `Bearer ${token}`
+        checkExpiration(localStorage.getItem('access_token'),handleLogout);
+        if (localStorage.getItem('access_token')) {
+            try {
+                const isConfirmed = window.confirm("Are you sure you want to delete this users?");
+                if (!isConfirmed) {
+                    return; // Do nothing if the user cancels
                 }
-            });
-            if (!response.ok) {
-                throw new Error('Network response was not ok');
+
+                const BASE_URL = process.env.REACT_APP_BASE_URL ||  'http://localhost:8080';
+                const token = localStorage.getItem('access_token');
+
+                const response = await fetch(`${BASE_URL}/users/delete-admin?email=${email}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                        'Authorization': `Bearer ${token}`
+                    }
+                });
+                if (!response.ok) {
+                    throw new Error('Network response was not ok');
+                }
+                setUsers(currentAdmins => currentAdmins.filter(users => users.email !== email));
             }
-            setUsers(currentAdmins => currentAdmins.filter(users => users.email !== email));
-        }
-        catch (error) {
-            console.error("Error fetching admin data:", error);
+            catch (error) {
+                console.error("Error fetching admin data:", error);
+            }
         }
     };
